added spell cast registry and dynamic spell cast asignment

+++ src/main/java/at/kaindorf/arcane_conjouring/item/wand/addon/spell/BlockBreakSpellCast.java
@@ -4,12 +4,12 @@ import at.kaindorf.arcane_conjouring.item.wand.addon.CastingTarget;
 import net.minecraft.core.BlockPos;
 import net.minecraft.world.item.Tier;
 import net.minecraft.world.level.Level;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockState;
 import net.minecraftforge.common.TierSortingRegistry;
 
 public class BlockBreakSpellCast extends SpellCast{
-
-    Tier tier;
-    public BlockBreakSpellCast(Tier tier) {
+    public BlockBreakSpellCast() {
         super(2);
     }
 
@@ -20,8 +20,7 @@ public class BlockBreakSpellCast extends SpellCast{
 
         BlockPos block = target.getBlock();
         Level level = target.getLevel();
-        if (block == null || level == null) return;
-
+        if (block == null || level == null || level.getBlockState(block).equals(Blocks.BEDROCK.defaultBlockState())) return;
         level.destroyBlock(block, true);
     }
 }
